requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow>=10.0
django-storages[s3]>=1.14
boto3>=1.34
gunicorn>=21.2
whitenoise>=6.6
python-dotenv>=1.0
dj-database-url>=2.1
qrcode[pil]>=7.4
python-dateutil>=2.8
